Fixed Added Mock test and added Service to add weather data.

+++ src/main/java/ee/taltech/testing/courseproject/DTO/WeatherDTO.java
@@ -12,7 +12,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class WeatherDTO {
-    private int temp;
+    private double temp;
     private int pressure;
     private int humidity;
 }
